requirements.txt
langchain
langchain-ollama
langchain-openai
langchain_core
langchain_chroma
langchain_postgres
python-dotenv
langchain_community
openpyxl
pandas
numpy
pgvector
psycopg2
openai
unsloth
optuna
optuna-dashboard
datasets
unsloth 
vllm
rouge-score
sacrebleu
evaluate 
nltk
bert_score